+++ cmd/deploy.go
@@ -31,8 +31,6 @@ const (
 	REGISTRY_FLAG                     = "registry"
 	STORAGE_CLASS_FLAG                = "storage-class"
 	LOW_RESOURCES_FLAG                = "low-resources"
-	ENABLE_CUSTOM_METRICS_FLAG        = "custom-metrics"
-	ENABLE_KUBE_STATE_METRICS_FLAG    = "kube-state-metrics"
 	STORE_ISSUES_LOGS_ONLY_FLAG       = "store-issues-logs-only"
 	STORE_ISSUES_LOGS_ONLY_KEY        = "storeIssuesLogsOnly"
 	CHART_NAME                        = "groundcover/groundcover"
@@ -46,8 +44,6 @@ const (
 	CLUSTER_URL_FORMAT                = "%s/?clusterId=%s&viewType=Overview"
 	QUAY_REGISTRY_PRESET_PATH         = "presets/quay.yaml"
 	AGENT_KERNEL_5_11_PRESET_PATH     = "presets/agent/kernel-5-11.yaml"
-	CUSTOM_METRICS_PRESET_PATH        = "presets/backend/custom-metrics.yaml"
-	KUBE_STATE_METRICS_PRESET_PATH    = "presets/backend/kube-state-metrics.yaml"
 	STORAGE_CLASS_TEMPLATE_PATH       = "templates/backend/storage-class.yaml"
 	WAIT_FOR_GET_LATEST_CHART_FORMAT  = "Waiting for downloading latest chart to complete"
 	WAIT_FOR_GET_LATEST_CHART_SUCCESS = "Downloading latest chart completed successfully"
@@ -86,12 +82,6 @@ func init() {
 
 	DeployCmd.PersistentFlags().Bool(STORE_ISSUES_LOGS_ONLY_FLAG, false, "store issues logs only")
 	viper.BindPFlag(STORE_ISSUES_LOGS_ONLY_FLAG, DeployCmd.PersistentFlags().Lookup(STORE_ISSUES_LOGS_ONLY_FLAG))
-
-	DeployCmd.PersistentFlags().Bool(ENABLE_CUSTOM_METRICS_FLAG, true, "enable custom metrics scraping")
-	viper.BindPFlag(ENABLE_CUSTOM_METRICS_FLAG, DeployCmd.PersistentFlags().Lookup(ENABLE_CUSTOM_METRICS_FLAG))
-
-	DeployCmd.PersistentFlags().Bool(ENABLE_KUBE_STATE_METRICS_FLAG, true, "enable kube state metrics deployment")
-	viper.BindPFlag(ENABLE_KUBE_STATE_METRICS_FLAG, DeployCmd.PersistentFlags().Lookup(ENABLE_KUBE_STATE_METRICS_FLAG))
 
 	DeployCmd.PersistentFlags().String(COMMIT_HASH_KEY_NAME_FLAG, "", "the annotation/label key name that contains the app git commit hash")
 	viper.BindPFlag(COMMIT_HASH_KEY_NAME_FLAG, DeployCmd.PersistentFlags().Lookup(COMMIT_HASH_KEY_NAME_FLAG))
@@ -605,16 +595,6 @@ func generateChartValues(chartValues map[string]interface{}, apiKey, installatio
 
 	if viper.GetString(REGISTRY_FLAG) == "quay" {
 		overridePaths = append(overridePaths, QUAY_REGISTRY_PRESET_PATH)
-	}
-
-	enableCustomMetrics := viper.GetBool(ENABLE_CUSTOM_METRICS_FLAG)
-	if enableCustomMetrics {
-		overridePaths = append(overridePaths, CUSTOM_METRICS_PRESET_PATH)
-	}
-
-	enableKubeStateMetrics := viper.GetBool(ENABLE_KUBE_STATE_METRICS_FLAG)
-	if enableKubeStateMetrics {
-		overridePaths = append(overridePaths, KUBE_STATE_METRICS_PRESET_PATH)
 	}
 
 	if semver.MustParseRange(">=5.11.0")(nodesReport.MaximalKernelVersion()) {